requirements.txt
dash==2.17.1
dash_ag_grid==31.2.0
Langchain==0.2.6
langchain_core==0.2.10
Langchain_groq==0.1.5
langchain_groq==0.1.5
pandas==2.2.2
plotly==5.22.0
python-dotenv==1.0.1
